lab05/requirements.txt
Flask==2.1.1
Werkzeug==2.2.2
psycopg2-binary==2.9.3
